+++ client/src/App.tsx
@@ -1,29 +1,42 @@
-import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
-import Page from "./components/Page";
-import {
-    Home,
-    Prices,
-    Profile
-} from "./pages"
+import { BrowserRouter as Router, Route, Switch } from "react-router-dom"
+import { useState } from "react"
+import Page from "./components/Page"
+import { Home, Profile } from "./pages"
+import { Landing } from "./pages/Landing"
+import { SignIn } from "./pages/SignIn"
+import { SignUp } from "./pages/SignUp"
 
 function App() {
+    const [signInData, setSignInData] = useState({email: "", password: ""})
+    const [signUpData, setSignUpData] = useState({name: "", email: "", password: ""})
+
     return (
         <Router>
             <Switch>
-                <Route path="/prices">
+                <Route path="/sign-in">
+                    <SignIn
+                        data={signInData}
+                        onDataChange={setSignInData}
+                        onSubmit={event => event.preventDefault()}/>
+                </Route>
+                <Route path="/sign-up">
+                    <SignUp
+                        data={signUpData}
+                        onDataChange={setSignUpData}
+                        onSubmit={event => event.preventDefault()}/>
+                </Route>
+                <Route path="/home">
                     <Page>
-                        <Prices />
+                        <Home/>
                     </Page>
                 </Route>
                 <Route path="/profile">
                     <Page>
-                        <Profile />
+                        <Profile/>
                     </Page>
                 </Route>
                 <Route path="/">
-                    <Page>
-                        <Home />
-                    </Page>
+                    <Landing/>
                 </Route>
             </Switch>
         </Router>
