+++ client/src/pages/SignIn.tsx
@@ -1,7 +1,7 @@
 import React, { Dispatch, FormEvent, SetStateAction } from "react"
 import { AppBar, Button, Container, makeStyles, TextField, Toolbar, Typography } from "@material-ui/core"
 import { handleStateChange } from "../handlers/Handlers"
-import { SignUpData } from "./data/SignUpData"
+import { SignInData } from "../components/data/SignInData"
 
 const useStyles = makeStyles(theme => ({
     navigation: {
@@ -16,6 +16,9 @@ const useStyles = makeStyles(theme => ({
         flexDirection: "column",
         alignItems: "center"
     },
+    navigationItem: {
+        margin: "4px"
+    },
     title: {
         marginTop: theme.spacing(1)
     },
@@ -27,37 +30,41 @@ const useStyles = makeStyles(theme => ({
     }
 }))
 
-export interface SignUpProps {
-    data: SignUpData
-    onDataChange: Dispatch<SetStateAction<SignUpData>>
+export interface SignInProps {
+    data: SignInData
+    onDataChange: Dispatch<SetStateAction<SignInData>>
     onSubmit: (event: FormEvent) => void
 }
 
-export function SignUp(props: SignUpProps) {
+export function SignIn(props: SignInProps) {
     const classes = useStyles()
 
     return (
         <>
             <AppBar position="static" color="transparent" elevation={0}>
                 <Toolbar className={classes.navigation}>
-                    <Button variant="text" color="primary">Home</Button>
-                    <Button variant="outlined" color="primary">Sign In</Button>
+                    <Button
+                        className={classes.navigationItem}
+                        href="/"
+                        variant="text"
+                        color="primary">
+                        Home
+                    </Button>
+                    <Button
+                        className={classes.navigationItem}
+                        href="/sign-up"
+                        variant="outlined"
+                        color="primary">
+                        Sign Up
+                    </Button>
                 </Toolbar>
             </AppBar>
 
             <Container className={classes.container}>
                 <Typography className={classes.title} component="h1" variant="h5">
-                    Sign Up
+                    Sign In
                 </Typography>
                 <form className={classes.form}>
-                    <TextField
-                        variant="outlined"
-                        margin="normal"
-                        fullWidth
-                        label="Name"
-                        value={props.data.name}
-                        onChange={handleStateChange("name", props.data, props.onDataChange)}
-                    />
                     <TextField
                         variant="outlined"
                         margin="normal"
@@ -80,8 +87,8 @@ export function SignUp(props: SignUpProps) {
                         variant="contained"
                         fullWidth
                         color="primary"
-                        onClick={props.onSubmit}>
-                        Sign Up
+                        href="/home">
+                        Sign In
                     </Button>
                 </form>
             </Container>
