+++ client/src/pages/SignUp.tsx
@@ -1,7 +1,7 @@
 import React, { Dispatch, FormEvent, SetStateAction } from "react"
 import { AppBar, Button, Container, makeStyles, TextField, Toolbar, Typography } from "@material-ui/core"
 import { handleStateChange } from "../handlers/Handlers"
-import { SignInData } from "./data/SignInData"
+import { SignUpData } from "../components/data/SignUpData"
 
 const useStyles = makeStyles(theme => ({
     navigation: {
@@ -16,6 +16,9 @@ const useStyles = makeStyles(theme => ({
         flexDirection: "column",
         alignItems: "center"
     },
+    navigationItem: {
+        margin: "4px"
+    },
     title: {
         marginTop: theme.spacing(1)
     },
@@ -27,29 +30,49 @@ const useStyles = makeStyles(theme => ({
     }
 }))
 
-export interface SignInProps {
-    data: SignInData
-    onDataChange: Dispatch<SetStateAction<SignInData>>
+export interface SignUpProps {
+    data: SignUpData
+    onDataChange: Dispatch<SetStateAction<SignUpData>>
     onSubmit: (event: FormEvent) => void
 }
 
-export function SignIn(props: SignInProps) {
+export function SignUp(props: SignUpProps) {
     const classes = useStyles()
 
     return (
         <>
             <AppBar position="static" color="transparent" elevation={0}>
                 <Toolbar className={classes.navigation}>
-                    <Button variant="text" color="primary">Home</Button>
-                    <Button variant="outlined" color="primary">Sign Up</Button>
+                    <Button
+                        className={classes.navigationItem}
+                        href="/"
+                        variant="text"
+                        color="primary">
+                        Home
+                    </Button>
+                    <Button
+                        className={classes.navigationItem}
+                        href="/sign-in"
+                        variant="outlined"
+                        color="primary">
+                        Sign In
+                    </Button>
                 </Toolbar>
             </AppBar>
 
             <Container className={classes.container}>
                 <Typography className={classes.title} component="h1" variant="h5">
-                    Sign In
+                    Sign Up
                 </Typography>
                 <form className={classes.form}>
+                    <TextField
+                        variant="outlined"
+                        margin="normal"
+                        fullWidth
+                        label="Name"
+                        value={props.data.name}
+                        onChange={handleStateChange("name", props.data, props.onDataChange)}
+                    />
                     <TextField
                         variant="outlined"
                         margin="normal"
@@ -72,8 +95,8 @@ export function SignIn(props: SignInProps) {
                         variant="contained"
                         fullWidth
                         color="primary"
-                        onClick={props.onSubmit}>
-                        Sign In
+                        href="/home">
+                        Sign Up
                     </Button>
                 </form>
             </Container>
